toke.c: C_ARRAY_END() doesn't work on a string
The code was using the macro C_ARRAY_END which doesn't apply to strings,
thus not giving the correct end to the string.  But no tests were
failing.  No new tests are added here, because the next commit will
change things so that tests would fail all over the place.