podcheck.t: Dont check verbatim length on a few pods
These pods have some very long lines that make sense to keep on a single
line, such as output from a program.  That means that someone viewing
them will either enlarge their window to view them unbroken or all is
lost anyway; there may be a few lines that could be shortened, but no
real value to do so.