threads/t/stack* Generalize larger min stack sizes
This changes the tests so that they will work on systems that require
larger minimum stack sizes than what we initially try.  The initial test
is skipped in this case, but the rest of the tests are adjusted from the
new information.

This was the least disruptive way I could come up with that didn't
assume things worked correctly before running the tests, yet still
adapted to a platform that has a larger minimum requirement than we
guess.

One could increase the initial guess to the maximum of any box found so
far (this would mean doubling it in this case), but what happens if a
larger case comes around?