Accept \p{nv=-0}
Some platforms already ignore the minus sign.  The Unicode standard says
the minus sign should be ignored always.

(But always doing so would cause it to not properly get the only
negative numeric value character in Unicode; I have submitted a ticket
to Unicode about this.  Nevertheless, it is in keeping with the letter
and spirit of the standard for the leading minus to be ignored for -0)