toke.c: Return to pre-existing gcb rules
As stated in the C comments added in this commit, it has always worked
to say s/mark1/mark2/ where the marks are literal characters.  This is
incorrect because the marks should be considered modifiers of the '/'
delimiters.  (The substitute should be changed to do something like
s/\x{foo/\x{bar}/.)  But that may not happen in time for 5.36.  So this
commit reverts to pre-existing behavior for the ASCII delimiters that
have always worked this way.  Non-ASCII delimiters will continue to work
the new way, which could break code expecting the old.  But use of
non-ASCII delimiters is pretty rare, and the users of that could be
expected to have more knowledge of Unicode and would be less likely to
do something contrary to the Unicode Standard than someone using the
typical ASCII delimiters.