threads/t/stack*: Add comment
I noticed that the first real test in each of these could succeed by
happenstance even if the implementation is broken.  If the numbers we
use happen to be the same as the system default, we'll get true even if
the system ignored the request to use our numbers.

Fixing the test is a chicken and egg problem, as to find out what the
default value is requires using the functions the test files are trying
to test.  A unit could be added to Configure, but that seems too much
work for just this purpose.

The comments added by this commit suggest using an out-of-the-ordinary
value instead of the power of 2 one that is currently used.  That would
minimize the likelihood of happenstance.