fix variable used for checking Dump of slightly under 1
We try to calculate a variable that is just slightly less than 1 to test
Dumping it. An epsilon for slightly above and slightly under are both
calculated, but mistakenly the former was used in both cases for the
actual Dump. On most systems, this won't matter since the values will
end up being the same. Some systems use floating point formats that it
will make a difference.