Add a way to skip utils/ files that need 64 bit ints
 # Failed test 83 - utils/zipdetails compiles at porting/utils.t line 85
 #      got "Integer overflow in hexadecimal number at utils/zipdetails line 1432.
 Integer overflow in hexadecimal number at utils/zipdetails line 2247.
 Integer overflow in hexadecimal number at utils/zipdetails line 2248.
 Integer overflow in hexadecimal number at utils/zipdetails line 2249.
 Integer overflow in hexadecimal number at utils/zipdetails line 2250.
 Integer overflow in hexadecimal number at utils/zipdetails line 2251.
 Integer overflow in hexadecimal number at utils/zipdetails line 2252.
 Integer overflow in hexadecimal number at utils/zipdetails line 2253.
 Integer overflow in hexadecimal number at utils/zipdetails line 2254.
 utils/zipdetails syntax OK
 "
 # expected "utils/zipdetails syntax OK
 "
 t/porting/utils .................................................. FAILED at test 83

This patch adds a way that utils of this form can exit early with a
warning containing the magic phrase "requires 64 bit integers", in
which case the compile test will be skipped.

I will push a corresponding patch to https://github.com/pmqs/zipdetails
which is the upstream repository for cpan/IO-Compress/bin/zipdetails
which is hosted at https://github.com/pmqs/IO-Compress

This problem seems to have started when IO-Compress 2.105 was merged
in 131aabd8eb09322f5fc1ad33ed607d45390ee16d.