Makefile.SH - reduce technical debt, use `runtests` for `make test_porting`
Currently `make test_porting` rolls its own `runtests` logic, which is
just technical debt waiting to cause trouble. It also means that the
porting tests do not actually get run *exactly* as they would through
`make test_harness` which seems like a bad thing generally. (Even if so
far nobody has noticed a problem.)

Some of our tests require special setup which is performed in
`runtests`. We should not unroll `runtests` just for the test_porting
target, even if /right now/ it happens to work most of the time. I say
most of the time because `make test_porting` won't do the same thing as
`make test_harness` would via `runtests` if the test is not executed
under a tty. It also doesn't fixup PWD on platforms that need it. (Not
sure if that is a problem, but it might be.) In the future we might add
additional guards or setup to runtests and then the `test_porting`
target would have to change.

Lastly, there is no reason to unroll `runtests` like this. It is quite
easy to do it right, so lets just do it right and not set a precedent
that is ok to bypass `runtests` in our Makefile infrastructure.