t/harness - use glob on arguments even on *nix
Historically in t/harness we used glob() on the arguments only on
Windows as its shell does not do wildcard expansion. However on *nix
build arguments passed into harness via the makefile are passed in via
the TEST_ARGS env var, where they are also NOT expanded by the shell.
The simple solution is to use glob on all arguments.

Historically we used the -re option, which does a similar task, for
`make test_reonly`, but this allows that target to be converted to use
the simpler and easier to understand globs. The -re and -nre arguments
are extracted from @ARGV prior to the glob() call so there is no
collision between the two. Frankly had we had this patch back in the day
I would have used globs when I created `make test_reonly` intead of
creating the -re option. But the -re option still has its uses so
this patch does not use it.

This means that it becomes trivial to create new specialized test make
targets and it makes it easier to do stuff like:

    make test_harness TEST_ARGS="-v re/pat*.t"

for instance, without requiring the use of the -re option to specify
the list of files.