Deparse: honour custom warnings categories
Perl's warning system is based around a bitmask string, with 2 bits per
warning category. There are a large number of predefined categories
(such as "syntax" etc) which map to known, fixed indexes within that
string.

However it is possible for modules to register custom warning categories
(indeed, the "constant" and "overload" warning categories are custom,
registered during "use constant" or "use overload").

Deparse, as originally written, masked out any custom categories when
deciding whether the warnings state had changed (and thus a 'no/use
warnings "foo"' needed emitting). This made the code easier as it didn't
have to cope with variable-length bitmasks.

However, after v5.27.5-387-g006c1a1dbd (which tried to handle the
problem of unused bits in the last byte of the bitmask), then depending
on how many pre-defined warning categories there were, as a side effect
caused Deparse to start handling anywhere between 0 and 3 custom warning
categories, depending on how many built-in categories there were in the
current release.

It just so happens that in the forthcoming release (5.36), the number of
built-in categories is a multiple of 4, causing Deparse to handle 0
custom categories. Which is causing some 't/TEST -deparse' tests to
fail.

To fix this, this commit goes the full way and makes Deparse honour
*all* custom warnings categories, both by removing the masking, and
by normalising bitmask lengths before comparing them.