update Porting/deparse-skips.txt
ext/B/t/optree_for.t is a new test script which gets upset when

    for (...) {}

gets deparsed as

    for (...) { (); }

which produces a subtly different optree.
So mark it as known failing for now.

t/op/not.t was a skip, but seems to have been succeeding for while, so
remove it from the skips file.