op/magic.t - make $SIG{ALRM} local test deal with inherited IGNORE'd signal
Signal ignores can be set up by a parent process and the child process
will inherit them. So in some situations when we test $SIG{ALRM} might
be "IGNORE" and not undef. So instead of just expecting undef, check
what it was before the localization, and then check it is still that
value afterwards.

I found this while running make test via rebase --exec, something like
this reduction (from Matthew Horsfall):

    $ git rebase --exec='perl -le"print \$SIG{ALRM}"' HEAD~1
    IGNORE

A similar case for a different signal would be this example (from Leon
Timmermans):

    $ nohup perl -E 'say $SIG{HUP}' 2>/dev/null | cat
    IGNORE