threads: stack.t: make Deparse-friendly
On non-threaded builds, it quits in the first BEGIN block because
threads aren't enabled.  This leaves an incomplete parse tree and thus
heavily stunted Deparse output - in particular the vars declared before
the BEGIN block are missing.

By moving the var declarations/initialisation to after the first BEGIN
block, it makes it compile again.