Socket: getnameinfo() behaves differently on z/OS
POSIX says at least one of the parameters must be non-NULL.  z/OS
requires both to be.

It would be better to have a Configure probe for this, but this is the
first non-conforming OS we have found, and at this stage, there aren't
likely to be any others that Perl might eventually be ported to.  Should
some come along, a probe could be added at that time.