Normalize: Use ppport.h
This commit changes Unicode::Normalize to use ppport.h.

All modules that convert from UTF-8 should be using the latest ppport.h,
so that they get the latest security checking available, which ppport.h
backports to 6.1.

This commit converts the module to use the utf8_to_uvchr_buf() method of
translation from UTF-8, instead of a method that allows one to pass
flags as to what code points to not accept.  Since the flags parameter
it was passing was zero, this is equivalent to utf8_to_uvchr_buf().