sv.c - add new bool related utility functions and macros
The new bool "type" does not have the usual complement of utility
functions and macros. It only has one encapsulating function, which is
perfectly reasonable for most use cases where one wants to test if an
SV* is a bool, but does a bit too much if one is working on a
serialization tool which is likely to want to unroll a nice chunk of the
logic.

The new type also lacks the usual cohort of utility functions to create
new bool SV's.

This patch adds the following functions:

    newSVbool(const bool bool_val)
    newSV_true()
    newSV_false()

    sv_set_true(SV *sv)
    sv_set_false(SV *sv)
    sv_set_bool(SV *sv, const bool bool_val)

And the following macros:

    SvIandPOK(sv)
    SvIandPOK_off(sv)
    SvIandPOK_on(sv)

The following three are intended very specifically for writing
serialization code like for Sereal, where it is reasonable to want to
unroll the logic contained in Perl_sv_isbool() and SvTRUE(). They are
documented as "you dont want to use this except under special
circumstances".

    SvBoolFlagsOK(sv)
    BOOL_INTERNALS_sv_isbool(sv)
    BOOL_INTERNALS_sv_isbool_true(sv)
    BOOL_INTERNALS_sv_isbool_false(sv)

SvBoolFlagsOK() is documented as being the same as SvIandPOK(),
but this abstracts the intent from the implementation.