Revert "Porting/makerel Make EBCDIC files writable"
This reverts commit ecef3a65d2c34ab6acbc8d1e8e17f5f41e307729.

This commit stems from an earlier mechanism for creating EBCDIC
releases, but no longer makes sense, leaving all files user-writable.
The intent was to do so just on EBCDIC builds, as at one time all files
were translated into EBCDIC on the destination box.  But nowadays, the
translation is done in makerel itself, before the creation of the tar.

After this commit, files are read-only except for those on the internal
exception list.

Spotted by Steve Hay.