regen/warnings.pl: uppercase 'global' lexical vars
various lexical vars like $tree, $def are populated early on, then their
values are used later in many places, including directly in subs.

Rename these vars to be uppercase and with more meaningful names, to
emphasise their globalness.

(Ideally they really ought to be local and passed as arguments to all
the subs that use them, but that's more work.)