autodoc.pl: A source is perlxs.pod
This was meant to be looked at, but wasn't.  It's skipping cpan, dist,
and ext, which it really should be, because we don't have much control
over those.

The comments indicated it expected to get generated files from
lib/ instead, but those wouldn't be in MANIFEST, so weren't getting
looked at.  And it's not a good idea to look for them anyway, as one
could get relicts.  Instead create an exception list of pods that could
have API elements that are documented.

The only current file on that list is what becomes lib/perlxs.pod