autodoc.pl: Include thread context in arg calculation
When several different functions are grouped together into one entry,
only the signature of the first one is output provided that all have the
same signature (except for the function name).

Until this commit, whether or not there was a hidden thread context
parameter was not considered in this calculation.  This could lead to
confusion.