autodoc.pl: Rmv redundant text
When different functions are grouped together in a single entry, some or
all of them may be deprecated or experimental.  Prior to this commit
each such was listed individually as a separate paragraph.  It is
shorter to say "All forms" if all are of the same ilk, or list just the
individual ones in a single sentence, otherwise.