lib/perl5db.pl: Add some S<> to pod
I added these to surround some C<...> that have spaces so that the
output construct is guaranteed to be on a single line.  These are mostly
commands to type, and it is clearer to the reader if they kept together.
I didn't bother with things that were guaranteed to be at the beginning
of a line, as they won't wrap anyway.