mathoms.c pod: Add perlapi section names
This adds section names of what section each non-adjacent function
should go into in perlapi, perlintern.

Sections are in effect until the next command changing it.  As functions
are added/removed from this file, they may inadvertently get placed in
the wrong api section by virtue of inheriting the previous section name.
Guard against that by specifying the proper section the pod for each
function that is not in the same pod /* comment */ as another function.

This change fixed a bug where one function was shown in the wrong
section in perlapi.