Use $section names for perlapi
perlapi is divided into different sections depending on the purpose of
the documented element.  One can specify the heading name in the apidoc
description, or one can specify a $heading instead that expands to the
heading name when perlapi is generated.  It is best to use the latter,
to allow the exact real heading name to be tweakable without changing
all the references to it.

This commit changes the few remaining references to the spelled out
heading name to the $heading form.