fix panic from eval {} inside /(?{...})/
GH #19680

Normally in code like

    eval {.... };

then even if the eval is the last statement in the file or sub, the
OP_LEAVETRY isn't the last op in the execution path: it's followed
by an OP_LEAVE or OP_LEAVESUB or whatever, which will be the op to
resume execution from after an exception is caught.

However, if the eval is the *last* thing within a regex code block:

    /(?{ ...; eval {....}; })/

then the op_next pointer of the OP_LEAVETRY op is actually NULL.

This confused S_docatch(), which wrongly assumed that a NULL
PL_restartop indicated that the caught exception should be rethrown,
popping execution back to the outer perl_run() call and hence leading to
the confused panic warning:

    "panic: restartop in perl_run"

The fix is to to separate out the "do we need to re-throw" test,
(PL_restartjmpenv != PL_top_env), from the "no more ops so no need to
re-enter the runops loop" test, (!PL_restartop).