perlapi: Deprecate 5 functions
and document two of them, which I had done before I realized they were
to be deprecated.

These functions are obsolete, and stemmed from a long-solved problem in
Windows.  There are a couple of uses on cpan of a couple of them.  But
people should just use PL_opnames[], for example, instead of the
function here that returns the address of that array.