doio.c: Rename formal parameters for clarity
This was named 'not_implicit' instead of 'explicit' because the latter
is a C++ keyword.

But a negative name increases the cognitive load of understanding, and
in this case resulted in some double negatives, which is worse.

'is_explicit' isn't a keyword and is a clearer name for this concept.