autodoc.pl: Properly handle a single column
The program attempts to make a table of some of the output for
compactness.  But if there is room for only a single column, it was
using the number of rows that it calculated were necessary should two
columns have worked; hence the output was only displaying the first half
of the rows in perlintern.pod.  What is needed is to do the loop a
partial final iteration, just enough to calculate the number of rows for
that final column.