regen/embed.pl: Allow perl_' prefix to names
Normally, the embed.fnc C flag requires a p flag to make sure that the
name starts with 'Perl_', to avoid polluting the caller's name space.
This commit makes an exception to that if the name already starts with
[Pp]erl.

This is because we have few old functions whose name starts with
'perl_', and hence they don't fit in with the newer scheme of things.

The meanings of the O and o flags in combination could have been
tweaked, but this is the simplest solution, namely:  If the name isn't
going to pollute, it's ok no matter what the flags say.