perlapi: Mark call_list as internal
This function is heavily dependent on non-public state information.

ExtUtils::ParseXS generates code that uses it to emulate the generated
XS boot code for a range of perls where this was problematic.  This
range ends before 5.22.  This code is #ifdef'd out on newer perls.

New XS code should not use this function.

There is one CPAN module that uses it independently, Parse::Perl, using
the same formulation that ExtUtils::ParseXS generates.