opset_len in Opcode.xs is a constant, so should be a static const
We can also remove many dMY_CXT declarations, as they are no longer needed
(and generate a now unused variable in threaded builds, hence compiler
warnings).

Previously it was part of the module's my_cxt_t, because it was a value
calculated from the interpreter variable PL_maxo. But PL_maxo itself is *not*
a variable - it was converted to a #define in Aug 2016 by commit
8d89205aa6324e7d:

    Remove PL_maxo

    We have an interpreter variable using memory, PL_maxo, which is
    defined to be the same as MAXO, a #defined constant.  As far as I can
    tell, it is never used in lvalue context, in core or on CPAN, except
    for the initialisation in intrpvar.h.

    It can simply be removed and replaced with a macro defined as equiva-
    lent to MAXO.

    It was added in this commit:

    commit 84ea024ac9cdf20f21223e686dddea82d5eceb4f
    Author: Perl 5 Porters <perl5-porters.nicoh.com>
    Date:   Tue Jan 2 23:21:55 1996 +0000

        perl 5.002beta1h patch: perl.h

        5.002beta1 attempted some memory optimizations, but unfortunately
        they can result in a memory leak problem.  This can be
        avoided by #define STRANGE_MALLOC.  I do that here until
        consensus is reached on a better strategy for handling the
        memory optimizations.

        Include maxo for the maximum number of operations (needed
        for the Safe extension).

    But apparently it is not needed for the Safe extension (tests pass
    without it).

What the author of that commit didn't realise was that Opcode had been split
out from Safe - the code in question is in this module not Safe.