GDBM_File: fix fetching mmapsize
Originally this came up from a Coverity complaint.

GDBM_SETMAXMAPSIZE accepts any of unsigned, unsigned long or size_t,
but GDBM_GETMAXMAPSIZE only accepts size_t.  Since this is the only
case that uses unsigned values we can safely switch it to size_t.

Unfortunately Coverity's analysis was pretty broken, it complained
about c_uv being uninitialised at the call to newSVuv, but its example
code flow went through the opt_dbname case, which sets vptr to != &c_uv
so the newSVuv() wouldn't execute anyway.

Before looking closely at the Coverity analysis and after finding the
bug fixed here I thought for a moment that Coverity had been tracing
into libgdbm, which could have caught the actual problem, but alas
that was not the case.

I expect this fix will not close CID 351943, and if it remains after
this is applied I'll close it as a false positive.

CID 351943.